增加用户语言，mongo查询修改

+++ message-server/src/main/java/com/wp/casino/messageserver/domain/RoomMessageContext.java
@@ -20,7 +20,7 @@ public class RoomMessageContext implements Serializable {
 
     private String plynickname;
 
-    private Integer plyguid;
+    private long plyguid;
 
     private String magic_id;
